requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-slugify>=8.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0